Moving SetMatIDFromLuaFunction and SetBndryIDFromLuaFunction into the lua layer
