+++ lua/framework/mesh/volume_mesher/volume_mesher_set_property.cc
@@ -1,9 +1,9 @@
 #include "framework/lua.h"
-
+#include "framework/mesh/mesh_continuum/mesh_continuum.h"
 #include "framework/mesh/mesh_handler/mesh_handler.h"
 #include "framework/mesh/surface_mesher/surface_mesher.h"
 #include "framework/mesh/volume_mesher/extruder/volmesher_extruder.h"
-
+#include "framework/utils/timer.h"
 #include "framework/runtime.h"
 #include "framework/mesh/logical_volume/logical_volume.h"
 
@@ -36,6 +36,225 @@ RegisterLuaFunctionAsIs(chiVolumeMesherSetKBAPartitioningPxPyPz);
 RegisterLuaFunctionAsIs(chiVolumeMesherSetKBACutsX);
 RegisterLuaFunctionAsIs(chiVolumeMesherSetKBACutsY);
 RegisterLuaFunctionAsIs(chiVolumeMesherSetKBACutsZ);
+
+using namespace opensn;
+
+namespace
+{
+
+/**
+ * Sets material id's using a lua function. The lua function is called with for each cell with 4
+ * arguments, the cell's centroid x,y,z values and the cell's current material id.
+ *
+ * The lua function's prototype should be:
+ * \code
+ * function LuaFuncName(x,y,z,id)
+ *   --stuff
+ * end
+ * \endcode
+ */
+void
+SetMatIDFromLuaFunction(const std::string& lua_fname)
+{
+  const std::string fname = "VolumeMesher::SetMatIDFromLuaFunction";
+
+  opensn::log.Log0Verbose1() << program_timer.GetTimeString()
+                             << " Setting material id from lua function.";
+
+  // Define console call
+  auto L = console.GetConsoleState();
+  auto CallLuaXYZFunction = [&L, &lua_fname, &fname](const Cell& cell)
+  {
+    // Load lua function
+    lua_getglobal(L, lua_fname.c_str());
+
+    // Error check lua function
+    if (not lua_isfunction(L, -1))
+      ChiLogicalError("Attempted to access lua-function, " + lua_fname +
+                      ", but it seems the function could not be retrieved.");
+
+    const auto& xyz = cell.centroid_;
+
+    // Push arguments
+    lua_pushnumber(L, xyz.x);
+    lua_pushnumber(L, xyz.y);
+    lua_pushnumber(L, xyz.z);
+    lua_pushinteger(L, cell.material_id_);
+
+    // Call lua function
+    // 4 arguments, 1 result (double), 0=original error object
+    int lua_return;
+    if (lua_pcall(L, 4, 1, 0) == 0)
+    {
+      LuaCheckNumberValue(fname, L, -1);
+      lua_return = lua_tointeger(L, -1);
+    }
+    else
+      ChiLogicalError("Attempted to call lua-function, " + lua_fname + ", but the call failed.");
+
+    lua_pop(L, 1); // pop the int, or error code
+
+    return lua_return;
+  };
+
+  // Get current mesh handler
+  auto& handler = GetCurrentHandler();
+
+  // Get back mesh
+  MeshContinuum& grid = *handler.GetGrid();
+
+  int local_num_cells_modified = 0;
+  for (auto& cell : grid.local_cells)
+  {
+    int new_matid = CallLuaXYZFunction(cell);
+
+    if (cell.material_id_ != new_matid)
+    {
+      cell.material_id_ = new_matid;
+      ++local_num_cells_modified;
+    }
+  } // for local cell
+
+  const auto& ghost_ids = grid.cells.GetGhostGlobalIDs();
+  for (uint64_t ghost_id : ghost_ids)
+  {
+    auto& cell = grid.cells[ghost_id];
+    int new_matid = CallLuaXYZFunction(cell);
+
+    if (cell.material_id_ != new_matid)
+    {
+      cell.material_id_ = new_matid;
+      ++local_num_cells_modified;
+    }
+  } // for ghost cell id
+
+  int globl_num_cells_modified;
+  MPI_Allreduce(
+    &local_num_cells_modified, &globl_num_cells_modified, 1, MPI_INT, MPI_SUM, mpi.comm);
+
+  opensn::log.Log0Verbose1() << program_timer.GetTimeString()
+                             << " Done setting material id from lua function. "
+                             << "Number of cells modified = " << globl_num_cells_modified << ".";
+}
+
+/**Sets boundary id's using a lua function. The lua function is called for each boundary face
+ * with 7 arguments, the face's centroid x,y,z values, the face's normal x,y,z values and the
+ * face's current boundary id. The function must return a new_bndry_name (string).
+ *
+ * The lua function's prototype should be:
+ * \code
+ * function LuaFuncName(x,y,z,nx,ny,nz,id)
+ * --stuff
+ * end
+ * \endcode
+ */
+void
+SetBndryIDFromLuaFunction(const std::string& lua_fname)
+{
+  const std::string fname = "VolumeMesher::SetBndryIDFromLuaFunction";
+
+  ChiLogicalErrorIf(opensn::mpi.process_count != 1, "Can for now only be used in serial.");
+
+  opensn::log.Log0Verbose1() << program_timer.GetTimeString()
+                             << " Setting boundary id from lua function.";
+
+  // Define console call
+  auto L = console.GetConsoleState();
+  auto CallLuaXYZFunction = [&L, &lua_fname, &fname](const CellFace& face)
+  {
+    // Load lua function
+    lua_getglobal(L, lua_fname.c_str());
+
+    // Error check lua function
+    ChiLogicalErrorIf(not lua_isfunction(L, -1),
+                      "Attempted to access lua-function, " + lua_fname +
+                        ", but it seems the function could not be retrieved.");
+
+    const auto& xyz = face.centroid_;
+    const auto& n = face.normal_;
+
+    // Push arguments
+    lua_pushnumber(L, xyz.x);
+    lua_pushnumber(L, xyz.y);
+    lua_pushnumber(L, xyz.z);
+    lua_pushnumber(L, n.x);
+    lua_pushnumber(L, n.y);
+    lua_pushnumber(L, n.z);
+    lua_pushinteger(L, static_cast<lua_Integer>(face.neighbor_id_));
+
+    // Call lua function
+    // 7 arguments, 1 result (string), 0=original error object
+    std::string lua_return_bname;
+    if (lua_pcall(L, 7, 1, 0) == 0)
+    {
+      LuaCheckNumberValue(fname, L, -1);
+      LuaCheckStringValue(fname, L, -2);
+      lua_return_bname = lua_tostring(L, -1);
+    }
+    else
+      ChiLogicalError("Attempted to call lua-function, " + lua_fname + ", but the call failed.");
+
+    lua_pop(L, 1); // pop the string, or error code
+
+    return lua_return_bname;
+  };
+
+  // Get current mesh handler
+  auto& handler = GetCurrentHandler();
+
+  // Get back mesh
+  MeshContinuum& grid = *handler.GetGrid();
+
+  // Check if name already has id
+  auto& grid_bndry_id_map = grid.GetBoundaryIDMap();
+
+  int local_num_faces_modified = 0;
+  for (auto& cell : grid.local_cells)
+    for (auto& face : cell.faces_)
+      if (not face.has_neighbor_)
+      {
+        const std::string bndry_name = CallLuaXYZFunction(face);
+        const uint64_t bndry_id = grid.MakeBoundaryID(bndry_name);
+
+        if (face.neighbor_id_ != bndry_id)
+        {
+          face.neighbor_id_ = bndry_id;
+          ++local_num_faces_modified;
+
+          if (grid_bndry_id_map.count(bndry_id) == 0) grid_bndry_id_map[bndry_id] = bndry_name;
+        }
+      } // for bndry face
+
+  const auto& ghost_ids = grid.cells.GetGhostGlobalIDs();
+  for (uint64_t ghost_id : ghost_ids)
+  {
+    auto& cell = grid.cells[ghost_id];
+    for (auto& face : cell.faces_)
+      if (not face.has_neighbor_)
+      {
+        const std::string bndry_name = CallLuaXYZFunction(face);
+        const uint64_t bndry_id = grid.MakeBoundaryID(bndry_name);
+
+        if (face.neighbor_id_ != bndry_id)
+        {
+          face.neighbor_id_ = bndry_id;
+          ++local_num_faces_modified;
+
+          if (grid_bndry_id_map.count(bndry_id) == 0) grid_bndry_id_map[bndry_id] = bndry_name;
+        }
+      } // for bndry face
+  }     // for ghost cell id
+
+  int globl_num_faces_modified;
+  MPI_Allreduce(
+    &local_num_faces_modified, &globl_num_faces_modified, 1, MPI_INT, MPI_SUM, mpi.comm);
+
+  opensn::log.Log0Verbose1() << program_timer.GetTimeString()
+                             << " Done setting boundary id from lua function. "
+                             << "Number of cells modified = " << globl_num_faces_modified << ".";
+}
+
+} // namespace
 
 int
 chiVolumeMesherSetProperty(lua_State* L)
@@ -178,9 +397,7 @@ chiVolumeMesherSetProperty(lua_State* L)
     int sense = true;
     if (num_args == 4) sense = lua_toboolean(L, 4);
 
-    if (bndry_name.empty())
-      throw std::invalid_argument(fname + ": argument 3 must not be"
-                                          "an empty string.");
+    ChiLogicalErrorIf(bndry_name.empty(), "argument 3 must not be an empty string.");
 
     const auto& log_vol =
       opensn::Chi::GetStackItem<LogicalVolume>(opensn::Chi::object_stack, volume_hndl, fname);
@@ -193,7 +410,7 @@ chiVolumeMesherSetProperty(lua_State* L)
 
     const std::string lua_fname = lua_tostring(L, 2);
 
-    VolumeMesher::SetMatIDFromLuaFunction(lua_fname);
+    SetMatIDFromLuaFunction(lua_fname);
   }
   else if (property_index == VMP::BNDRYID_FROM_LUA_FUNCTION)
   {
@@ -201,7 +418,7 @@ chiVolumeMesherSetProperty(lua_State* L)
 
     const std::string lua_fname = lua_tostring(L, 2);
 
-    VolumeMesher::SetBndryIDFromLuaFunction(lua_fname);
+    SetBndryIDFromLuaFunction(lua_fname);
   }
   else
   {
